MLIBZ-2040: Subclassing File was not possible because mapping() must be open and not just public
